% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evyth.R
\docType{package}
\name{evyth}
\alias{evyth}
\title{\code{evyth} package}
\description{
Caja de Herramientas para el procesamiento de la Encuesta de Turismo y Viaje de los Hogares (EVyTH)
See the README on
\href{https://github.com/dnme-minturdep/evyth/blob/main/README.md}{Github}
}
