% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_microdatos.R
\name{descargar_microdatos}
\alias{descargar_microdatos}
\title{Descarga de Bases de la Encuesta de Viajes y Turismo de los Hogares (EVyTH)}
\usage{
descargar_microdatos(anio = 2018, trimestre = NA, variables = "all")
}
\arguments{
\item{anio}{un integer o vector de integers a partir de 2012}

\item{trimestre}{un integer o vector de integers con el numero de trim: 1,2,3,4}

\item{variables}{un vector de characters. variables a seleccionar. Default='all' trae todas las variables}
}
\description{
Funcion que descarga bases de la Encuesta de Viajes y Turismo de lo Hogares
}
\examples{

base_individual <- descargar_microdatos(anio = 2018:2019,
                                      trimestre = 1,
                                      variables = c('anio','tipo_visitante','pondera'))


}
