% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escribir_base_microdato.R
\name{escribir_base_microdato}
\alias{escribir_base_microdato}
\title{Title}
\usage{
escribir_base_microdato(archivo = NULL, formato = "csv")
}
\arguments{
\item{formato}{}
}
\description{
Title
}
