% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_ipc.R
\name{obtener_ipc}
\alias{obtener_ipc}
\title{Descarga de tabla con serie de Indices IPC Cobertura Nacional}
\usage{
obtener_ipc(mes, anio)
}
\arguments{
\item{mes}{Variable numerica que indica el mes de publicacion del ipc que se busca}

\item{anio}{Variable numerica que indica el anio de publicacion del ipc que se busca}
}
\value{
dataframe con serie del indice de IPC Cobertura Nacional
}
\description{
Descarga de tabla con serie de Indices IPC Cobertura Nacional
}
\examples{
\dontrun{
obtener_ipc(mes = "10", anio = "2022")
}
}
